Moved import handling responsibility to a separate module.
Previously the Parser.Program module would handle the import system,
which violated the Single Responsibility Principle™. This made the
code very messy so I created the Parser.ModuleResolver module.
