--- test/parsing/Foo/Bar.zo
@@ -1,13 +1,0 @@
-module Bar {bar, Test{ (::) }}
-
-let bar x := x
-
-let-infix :: 7 right
-
-#:
-
-    #:
-        #:
-            NESTED MULTILINE COMMENT :#
-    :#
-:#
