+++ test/parsing/constructor_operators/Another.zo
@@ -1,0 +1,19 @@
+#: should export the operator
+module Another {Type {(:+)}}
+module Another {Type {Constructor, (:+)}}
+module Another
+module Another {Type {(:+), Constructor}}
+:#
+
+module Another {Type} # does not pass
+
+#: should not export the operator
+module Another {}
+module Another {Type {}}
+module Another {Type {Constructor}}
+:#
+
+let-infix :+ 6 right
+type Type a : * -> * with
+    case (:+) (Type a) a
+    case Constructor
