--- test/parsing/modules/Another.zo
@@ -1,5 +1,0 @@
-module Another
-
-let test () := 42
-
-let-infix ++ 5 right
