--- test/parsing/modules/DefaultOperatorTest.zo
@@ -1,9 +1,0 @@
-module DefaultOperatorTest
-
-
-let _ := x * y * z
-#          ^ first definition of "*"
-# let-infix * 5 right
-# ^ illegal redeclaration of the operator (*)
-
-let _ := 42
