--- test/parsing/other/Test.zo
@@ -1,1 +1,0 @@
-module Test {}
